requirements.txt
pyrogram
aiohttp
python-dotenv